# storage/api-docs/api-docs.yaml
...
openapi: "3.0.0"
info:
  title: "Upstart API"
  version: "1.0.0"
paths:
...
  /api/v1/{trainer_id}/courses:
    get:
      tags:
        - "General"
      summary: "Get trainer with his courses"
      description: "Get trainer with his courses"
      operationId: "TrainerCourses"
      parameters:
        - name: trainer_id
          in: path
          schema:
            type: integer
            example: 1000
      responses:
        "200":
          description: "Trainer courses"
        "403" :
          description: "User not found/Permission denied!"
  /api/v1/get-notifications/{id}:
    get:
      tags:
        - "General"
      summary: "Get Notifications"
      description: This request is  for logged in users. In  the request headers  need set user api_token as the bearer token.
      operationId: "Notification"
      parameters:
        - name: id
          in: path
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        "403":
          description: "Forbidden Error"
      security:
        - bearerAuth: [ ]
...
  /api/v1/categories:
    get:
      tags:
        - "Category"
      summary: "Get Categories"
      operationId: "Category"
      parameters:
        - name: "language_code"
          in: "query"
          schema:
            type: "string"
            example: en
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "array"
                items:
                  properties:
                    id:
                      type: "integer"
                    parent_id:
                      type: "integer"
                    ordering:
                      type: "integer"
                    title:
                      type: "string"
                    icon:
                      type: "string"
                    child:
                      type: "array"
                      items:
                        properties:
                          id:
                            type: "integer"
                          parent_id:
                            type: "integer"
                          ordering:
                            type: "integer"
                          title:
                            type: "string"
                          icon:
                            type: "string"
                        type: "object"
                  type: "object"
        "401":
          description: "Unauthenticated"
...
  /api/v1/courses:
    get:
      tags:
        - "Course"
      summary: "Get Courses"
      operationId: "Courses"
      parameters:
        - name: "category_id"
          in: "query"
          schema:
            type: "string"
            example: 1
        - name: "page"
          in: "query"
          schema:
            type: "integer"
            example: 1

        - name: "limit"
          in: "query"
          schema:
            type: "integer"
            example: 1
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Course'
        "401":
          description: "Unauthenticated"
...
  /api/v1/courses/{course_id}:
    get:
      tags:
        - "Course"
      summary: "Get Course by ID"
      operationId: "CategoryByID"
      parameters:
        - name: "language_code"
          in: path
          schema:
            type: "string"
            example: hy
        - name: course_id
          in: path
          schema:
            type: integer
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 1
                  category_id:
                    type: integer
                    format: int64
                    example: 1
                  trainer_id:
                    type: integer
                    format: int64
                    example: 1
                  title:
                    type: string
                    example: title course
                  sub_title:
                    type: string
                    example: sub title course
                  description:
                    type: string
                    example: description course
                  language:
                    type: integer
                    example: 1
                  status:
                    type: integer
                    example: 1
                  price:
                    type: integer
                    example: 1.1
                  max_participants:
                    type: integer
                    example: 1
                  level:
                    type: number
                    example: 1.1
                  currency:
                    type: string
                  promo_video:
                    type: string
                    format: binary
                  cover_image:
                    type: string
                    format: binary
                  lessons:
                    type: array
                    xml:
                      wrapped: true
                    items:
                      type: object
                      properties:
                        title:
                          type: string
                          example: Laravel
                        course_id:
                          type: integer
                          example: 1
                        duration:
                          type: integer
                          example: 1
                        start_time:
                          type: string
                          example: 2022-10-10 10:12:00
                  trainer:
                    type: object
                    properties:
                      first_name:
                        type: string
                        example: Trainer
                      last_name:
                        type: string
                        example: Trainer
                      bio:
                        type: string
                        example: bio
                      avatar:
                        type: string
        "403":
          description: "Forbidden Error"
      security:
        - bearerAuth: [ ]
  /api/v1/courses/create:
    post:
      tags:
        - "Course"
      summary: "Create Course"
      operationId: "CreateCourse"
      description: Create Course.This request is   for Trainer and Trainer Center  users. In  the request headers  need set user api_token as the bearer token.
      parameters:
        - name: "language_code"
          in: "query"
          schema:
            type: "string"
            example: hy
        - name: "type"
          in: "query"
          schema:
            type: "integer"
            example: 1
      responses:
        "200":
          description: "Course created successfully"

        "403":
          description: "Forbidden Error"
      security:
        - bearerAuth: [ ]
  /api/v1/courses/update:
    post:
      tags:
        - "Course"
      summary: "Update Course"
      operationId: "UpdateCourse"
      parameters:
        - name: "language_code"
          in: "query"
          schema:
            type: "string"
            example: hy
      requestBody:
        description: Update Course.This request is   for Trainer and Trainer Center  users. In  the request headers  need set user api_token as the bearer token.
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                  example: 1
                category_id:
                  type: integer
                  format: int64
                  example: 1
                trainer_id:
                  type: integer
                  example: 1
                title:
                  type: string
                  example: title course
                sub_title:
                  type: string
                  example: sub title course
                description:
                  type: string
                  example: description course
                language:
                  type: integer
                  example: 1
                requirements:
                  type: string
                  example: "['HTML', 'CSS']"
                will_learn:
                  type: string
                  example: "['JS', 'PHP']"
                status:
                  type: integer
                  example: 1
                price:
                  type: integer
                  example: 1.1
                max_participants:
                  type: integer
                  example: 1
                level:
                  type: number
                  example: 1
                currency:
                  type: string
                promo_video:
                  type: string
                  format: binary
                type:
                  type: integer
                  example: 1
                cover_image:
                  type: string
                  format: binary
                address:
                  type: string
                  format: binary
                lessons:
                  type: array
                  xml:
                    wrapped: true
                  items:
                    type: object
                    properties:
                      title:
                        type: string
                        example: Laravel
                      course_id:
                        type: integer
                        example: 1
                      duration:
                        type: integer
                        example: 1
                      start_time:
                        type: string
                        example: 2022-10-10 10:12:00
                trainer:
                  type: object
                  properties:
                    first_name:
                      type: string
                      example: Trainer
                    last_name:
                      type: string
                      example: Trainer
                    bio:
                      type: string
                      example: bio
                    avatar:
                      type: string
          application/xml:
            schema:
              $ref: '#/components/schemas/Course'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Course'
        required: true
      responses:
        "200":
          description: "Course updated successfully"

        "403":
          description: "Forbidden Error"
      security:
        - bearerAuth: [ ]
...
  /api/v1/courses/update-section:
    post:
      tags:
        - "Course"
      summary: "Update Course"
      operationId: "updateSection"
      description: Update Section.This request is   for Trainer and Trainer Center  users. In  the request headers  need set user api_token as the bearer token.
      parameters:
        - name: "language_code"
          in: "query"
          schema:
            type: "string"
            example: hy
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                title:
                  type: string

      responses:
        "200":
          description: "Course created successfully"

        "403":
          description: "Forbidden Error"
      security:
        - bearerAuth: [ ]
...
  /api/v1/moderators:
    get:
      tags:
        - "Moderator"
      summary: "Get Moderators"
      operationId: "Moderator"
      parameters:
        - name: "language_code"
          in: "query"
          schema:
            type: "string"
            example: hy
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: '#/components/schemas/Moderator'
        "401":
          description: "Unauthenticated"
      security:
        - bearerAuth: [ ]
...
components:
  schemas:
...
    Moderator:
      type: object
      properties:
        first_name:
          type: "string"
        last_name:
          type: "string"
        email:
          type: "string"
        password:
          type: "string"
        company_name:
          type: "string"
        role_id:
          type: integer
          default: 2
        tax_identity_number:
          type: "integer"
    Notification:
      type: object
      properties:
        id:
          type: "integer"
        title:
          type: "string"
        message:
          type: "string"
        type:
          type: "string"
        status:
          type: "integer"
        created_at:
          type: "string"
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT